Print a warning when creating new Analytics functions that their activation will be slow (#174)
Print a warning when creating new Analytics functions that their activation will be slow.

For most of beta, Analytics functions that are freshly created by the CLI will take up to six hours to start receiving events. This is sad, but unavoidable. In the mean time we'll print a loud (bold, yellow) message to warn users that they should come back to their newly created Analytics function in several hours' time.