Add pre & post-deploy hoooks to firebase deploy.
Each feature's object firebase.json can now include a script with "predeploy" or "postdeploy". The commands are run as a script in the system's default shell. Output is directly piped to the CLI's STDOUT and STDERR. STDIN is also piped to allow for interactive commands (e.g. one could imagine a script that requires a git tag to be named for all deploys). Any non-zero exit code or exit due to signal interruption will be considered a failures. Failures in predeploy are fatal, but there does not seem to be any reason to consider a failure in postdeploy scripts fatal.

Thoughts for reviewers/future work:

1. The CWD is currently the project root. Should it be a feature directory? This seems best for Functions but doesn't make sense for Hosting, Storage, etc.
2. What environment variables should we set for scripts? Maybe this could make (1) better.
3. This API was approved before the functions shell existed. Do we need another hook for the shell? E.g. compiling TS before emulating?
4. Should postdeploy always run, even on deploy failure? If so, should there be an environment variable for the deploy status? If not, should there be another lifecycle hook for deploy failures?