Add debug statements to functions emulation (#558)
We've received a handful of public issues regarding the emulator freezing on startup. There are some heavyweight operations that might hang and it's not clear which is failing. This will add a bit more info to the verbose log about what the CLI is doing up to the point of failure.

Also, it was probably incorrect to use .catch in the main trigger flow since it would also catch errors starting the emulator and blame trigger parsing. Moved that error decoration to be inside the nested Promise.