Add MultiDB support for database:* commands
This change allows a `--instance` option that lets you change the
instance you're using. I went with `--instance` because it makes the
rest of the code a no-op change and fits in well :)

This also does validation against the databases for your project,
recommending others if an invalid one is selected.