Add functions deploy functional test to travis
Adding functions test to run in a separate stage. We can probably combine it with hosting in the same stage - from what I've seen they both create separate tmp directories and work from there. Choosing separate stages for now for less risk. Also, perhaps we can make the hosting script accept a project so that it can be run locally (I see it relies on an env var that gets set in travis - unclear which project it is using).