Fix JSON-breaking bug in serialExportUsers
When `serialExportUsers` function is called more than once,
it adds a redundant comma, which breaks the JSON format.

The bug is caused by a global variable `jsonSep` in an internal
`_writeUsersToFile` function. I assume that it was first introduced
to make JSON streamable.

I changed the code to make it generate `writeUsersToFile`
shared among recursive calls.