Fix `firebase apps:sdkconfig` for web apps
Fix the template name, and the templating of said template.

This PR fixes #2088, but creates worrying code duplication between
`src/hosting/implicitInit.ts` and `src/management/apps.ts`. It seems
like code in `src/hosting/implicitInit.ts` should use the code in
`src/management/apps.ts`, but that would break the caching in
(deprecated) `src/fetchWebSetup.ts`.

I'm willing to make necessary changes to tidy this up a little,
but the direction we want to take this is unclear to me.