Clarifies what `firebase init` can do for each "Firebase feature"
The current description of what `firebase init` can do for each of the listed "Firebase features" is not quite accurate or complete. Many of the descriptions actually describe _all_ the things that you can do with the CLI for that "Firebase feature" not just what `firebase init` can do.

Also, updated "Database" in this choices list to the more accurate "Realtime Database" (which is the product's actual name) and did some general wordsmithing and phrasing clarification/alignment.