Add backend namespace before rebasing codebase onto it. (#3276)
Add backend namespace before rebasing codebase onto it.

The Backend namespace defines types that are neither the __trigger
annotations nor the GCP API types. This follows Google best practices
for implementing services because it buffers the system from changes
in storage or API layers.

In this codebase, it will allow us to support new trigger discovery
techniques (e.g. a container contract HTTP API) and new versions
of APIs if they should ever be needed.

In this change I try out a new method for using args.Context; it's
based on the Go style of Context where a namespace can provide
a value using Context as a private implementation detail.
I'm personally a fan. WDYT?