Improve Pub/Sub function deployments (#3622)
Fixes some issues discovered when testing pub/sub function deployments manually.

GCFv2 has some behavior changes that we need to take into account in our CLI:

GCF no longer creates Pub/Sub topics automatically. We handle this in tasks.ts by manually creating any task we need. The underlying API is put so we are safe if the topic already exists.
GCF fails if pubsubTopic is provided in UpdateFunction irrespective of whether that topic even changed. We've moved the check for a change in topic to checkForInvalidChangeOfTrigger and no longer send the topic in the API call.