Stop supporting RuntimeConfig in v2 (#3630)
* Transition to generateUploadUrl

The GenerateUploadUrl API is now live for v2. After talking to
Diana, she said that deploys from uploads to different regions might
work today, but they don't like it and want to add validation to
break it in the future. At minimum it would potentially cause
features like GenerateDownloadUrl and cloud console source viewing
to break.

* Stop supporting RuntimeConfig in v2