Fix broken functions deploy e2e test. (#3635)
This isn't a test target that runs as part of CI; it's a useful one to run manually. Also including a small fix in the `functions/prepare.ts` script that I found after fixing the e2e test!

I'm removing the `unknownFilter` test because we now treat any filter as a valid one (Deploy code simply deploys nothing and reports "Deploy Complete"). This is subpar experience, but I've found fix to the issue to be a little more than few lines of change, and I'm not keen on fixing this bug right now.