Rename triggerType to signatureType, following functions-framework.
triggerType is no longer needed in FunctionsRuntimeBundle due to
environment variables refactor. In preparation for CloudEvent integration,
we rename signatureType to more apporpriate signatureType which can be
one of 'http' or 'event'. This follows the same convention used by the
GCF functions framework and prepares us to provide supporrt for 'cloudevent'
function signature.