Improve functions deploy time by polling more frequently.
Our pollers are built on Throttler Queues. These have a built in
backoff rate and GCF takes so long that we will invariably backoff
until we hit the maximum. When unspecified the maximum is 1m.

Looking at a single function deploy, we can say that we previously
wasted on average 30s sleeping (assume our second to last poll was
half way through the 0-60s range). With a new limit we should waste
only 5s polling, for an average savings of 25s. The savings will
be more dramatic for deploys of multiple functions.