Revive the runtime config loader in the emulator runtime. (#3974)
Most recent version of Firebase Functions SDK automatically picks up locally stored `.runtimeconfig.json` to populate the config entries.

However, due to a bug in some older version of the Function SDK, this process may fail. See the following issues for more detail:

* https://github.com/firebase/firebase-tools/issues/3793
* https://github.com/firebase/firebase-functions/issues/877

As a workaround, the emulator runtime will load the contents of the `.runtimeconfig.json` to the `CLOUD_RUNTIME_CONFIG` environment variable. In the future, we will bump up the minimum version of the Firebase Functions SDK required to run the functions emulator to v3.15.1 and get rid of this workaround.

Also - noticed a dead code that can be removed, so removed.