Load runtime config when loading triggers in the Functions Emulator (#4162)
Addresses https://github.com/firebase/firebase-tools/issues/4158.

I've tried to find a way to add a test for this usecase, but there isn't a framework where I can easily hook in a change like this without big investments. I'm leaning towards releasing this bugfix as is, and finding a less urgent time to add regression test.