Minor refactoring of function deploy (#4406)
Just moving around code in `prepare.ts` phase of the function deploy to make it easy to support multi-codebase deploys. This is a simple refactor that should introduces no change to deploy behavior. I've added annotation to each section to make it slightly easier to understand what'a going on in each section.

There are few light changes:

* We reduce scope of `backend.checkAvailability` check to just the endpoints that's being deployed (to allow deploy to proceed even if region of endpoint NOT being deployed is down).  
* We reduce scope `ensureServiceAgentRoles`  check to just the endpoints that's being deployed (to reduce IAM calls to bare minimum)