Add emulator support for multi codebase functions setup. (#4447)
Emulator will now load all functions discoverable in all codebases defined in user's repository.

The change ended up being fairly easy to implement, as we simply needed to `for` loop through the array of functions config and setup N emulatable backend (vs 1) for the FunctionsEmulator to load. This required a minor change on how triggers are invoked - the Functions Emulator will now take care of finding the correct emulatable backend associated w/ the trigger on invocation instead of taking it as an argument.

I've taken a separate PR to test this new feature of the emulator to make PR reviews slightly less burdensome:  https://github.com/firebase/firebase-tools/pull/4448.