Statically allocate socket path for each function emulator runtime process. (#4692)
This is first of several refactoring change we plan on making on the Functions Emulator to migrate bespoke IPC protocol between the emulator and the emulated functions.

Today, the invoking an HTTP function involves an intricate dance between the emulator and the emulated function:

1) On HTTP function invocation, Functions Emulator designates a socket path ([node doc](https://nodejs.org/api/net.html#net_identifying_paths_for_ipc_connections)). The chosen socket path is passed to the emulated process via IPC message sends.

2) Emulated function spins up an ephemeral express server, binds to the socket path, and responds to the HTTP requests.

This change proposes that instead of generating the socket path on _every_ request, we do it once at the emulated function process creation time.

FWIW, the sock path that was being "dynamically" generated at function invocation already generates the same socket path for each emulated function anyway. The change here simply generates a random socket path and saves it along with other metadata associated with the emulated function process.

There are obvious follow ups to this change (e.g. why do we spin up an ephemeral server on every function call?). This PR intentionally isn't tackling them to make the PR small. You can expect them on later PRs.