Refactor mechanism for health checking emulated functions process. (#4693)
Today, the Functions Emulator and Emulated Function process makes use of structured IPC so that the emulator can keep track of the state of the process. This is important to make sure only one request is handled at a time.

As another step towards replacing IPC-based comms to HTTP, we change the logic of health checking to see if an emulated HTTP function is ready to receive a request at given socket path. We now create a new `/__/health` route that is used to check if the emulated process's server is up and running.

I also added additional `/robots.txt` and `/favicon.ico` routes that always received 404 as suggested in the [Functions Framework contract](https://github.com/GoogleCloudPlatform/functions-framework/blob/master/README.md#url-space)