Fix "bug" where function timeout option did not apply to Functions Emulator. (#4745)
Several months ago, we started to refactor the Functions Emulator with the goal of the Emulator implement more closely resembling how GCF works.

One way the Emulator and GCF differed was how timeout was enforced - in GCF that work was done outside of the runtime process whereas in the emulated environment it was enforced by the runtime process itself.

Our work on redesigning the emulator still has much more work to do. It was premature of me to remove a well working feature - again, hubris got the best of me ("I'll add the feature back in just a week or two!".

Many users have spoken out - so I'm proposing here we bring back the timeout feature I ripped out few months ago. It'll be our responsibility to preserve this feature in the redesign-ed the Emulator.

Fixes https://github.com/firebase/firebase-tools/issues/2837