Refactor functions emulator to offload all process related logic to the worker class. (#4903)
Small refactor to move all process management logic from `FunctionsEmulator` to `FunctionsWorker`. I want to start moving all code pieces involved in management of a function process from the functions emulator (which I consider as control plane) to the worker (which I consider to be the data plane).

cc/ @joehan For visibility. Should not impact extensions emulator in any way.