Refactor runtime to spin up http server at startup instead of per invocation. (#4861)
Today, emulated HTTP function is invocation works as follows:

1) Functions Emulator (FE) receives HTTP request from client
2) FE sends an IPC to a runtime process assigned to the triggered HTTP function.
3) The runtime process spins up a new Express server and begins listening to the socket assigned via `PORT`.
4) FE spins while sending `__/health` request to the port assigned to the worker process. When the socket is ready, forwards the request to the assigned port.

It seems to me a bit wasteful to spin up an express server for every invocation of the function, so in this PR we spin up the HTTP server at startup. One downside of this change is that function timeout is no longer respected for HTTP function. There isn't a quick way to fix that in small chunk - we'll tackle that problem later in https://github.com/firebase/firebase-tools/pull/4867.