Refactor functions emulator to handle all functions through HTTP interface. (#4867)
Lets invoke Background functions exactly the same way HTTP functions are triggered.

Instead of sending event data/context over IPC, we instead have the Functions Emulator make HTTP request to the server the runtime process is listening over the socket path. Everything is just as before - we are talking over the unix domain sockets but now using HTTP protocol to communicate.

This gives us chance to move some of the logic for transforming event payload into the runtime which is what happens on production (by functions framework) and makes it possible to enforce timeout on per-request basis!