Goodbye invokeTrigger. Everything is http (unless in debug mode) (#4886)
This is it! It was all for this!

Now that all functions, regardless of signature, is invoked via an HTTP endpoint, we complete the refactor to remove all traces of IPC comm between the functions runtime and the functions emulator.

Instead of relying on inter-process messages from the runtime process to monitor the state of the function ("is it idle? is it processing request? is it done processing the request?") we simply let the Functions Emulator submit a request to the a worker associated with the trigger and infer status based on the state of the request.

The only way IPC is being used is when Functions Emulator is running in DEBUG mode. In debug mode, a single process is used to invoke all function triggers, and we use IPC to let the runtime process know which trigger should be called in the subsequent request. We'll tackle fixing this feature later - we can just let runtime for other language not support debug mode until we have a concrete plan on how we can support debug mode in the SDK itself.