Wire up pinTag (#5047)
Add an experiment called "pintags". When this is set, you can set a "pinTag" filed to true on functions or run rewrite configs. This will cause the CLI to ensure that the latest version of that run service has a tag (garbage collecting if we exceed 500) and then send that tag along with the API request for the version.

Once server-side changes are live, this will cause the hosting version to be forever locked to the Run Revision, which means rollbacks actually work. Was hoping this would fix preview channels as well, but we need to fix the fact that those functions are not deployed during preview channel deploys.