Enforce function timeout at worker level. (#5464)
Function timeout for emulator was enforced by the emulated function process.

Here we pull out the timeout enforcement at the worker level (analogous to "Data Plane" in production environment). When timeout is reached, the worker will abort the triggering request and kill the child process.

This refactor will remove the need to write code for enforcing timeout for emulated python functions.