feat(turtles) Implementing how we find turtles.
They are found on most continents, some islands and, in the case of sea
turtles, much of the ocean. Like other amniotes (reptiles, birds, and
mammals) they breathe air and do not lay eggs underwater, although many
species live in or around water.

--

This commit implements the Turtle discovery process.
We are giving Michael Scott a solid by showing him where to find the
turtles.