Fix bug where functions shell failed to invoke event triggered functions in debug mode. (#5609)
When running functions shell with `--inspect-functions` flag (aka debug mode), the shell command wasn't preparing the emulated process to execute the trigger invoked by the user.

Including a small refactoring to add a property to the FunctionsEmulator class to denote whether it's in debug mode or not for clarity.

I recognize that there isn't any test here. Sadly, there is no existing test suite for the functions shell command, and frankly I understand why - it's tricky to setup a test that requires interactive input. I'm intentionally not taking up the challenge to set up a test suite here since the bug is fairly small, but I do hope that we'll have time in the future to pay this tech debt (famous last words?).

Fixes https://github.com/firebase/firebase-tools/issues/5567