Fix bug where eagerly initializing UA failed function deployment that imported firebase-tools as a library.
We will lazily initialize the UA agent to prevent CF3 deploy from failing when building function source that uses firebase-tools as a library.

Not sure why this suddenly became a problem in nodejs18 - my guess is that the builder image for nodejs18 runtime have stricter permission on filesystem directories that configstore needs.