Inject GOOGLE_CLOUD_QUOTA_PROJECT environment variable during function discovery and emulation (#5985)
When deploying functions that makes use of GCP's usage-based (not resource-based) APIs , such as Vertex AI or ML Vision API, users may see an error like this:


```
// main.py
import vertexai
# Initialize the Vertex AI client
vertexai.init()

$ firebase deploy
Failed to load function definition from source: FirebaseError: Failed to parse build specification

Caused by:
403 Vertex AI API has not been used in project 563584335869 before or it is disabled. 
```

During function discovery, `vertex.init()` makes a call to the Vertex AI with the client credentials associated with Firebase CLI . In the case of usage-based APIs, the call is being made on the Firebase CLI project, not the users project, resulting in the error message like `been used in project 563584335869 before or it is disabled.` (project `563584335869` is the Firebase CLI GCP project).

Similar issue happens when running the function on the emulator.

To workaround the issue, we have to properly override the client project associated with the Vertex AI API call. There are few ways for doing this, and here we choose to leverage `GOOGLE_CLOUD_QUOTA_PROJECT` environment variable to correctly override the quota project to be the user project associated with the Firebase CLI session.