Disables KeepAlive timeout when debugger is attached to the functions emulator (#6069)
Node.js 19 introduced a change that sets `keepAlive` to true by default, with a default keep-alive duration of 5 seconds. This change broke our functions emulator behavior when a debugger was attached, as the Node.js HTTP server would consider the connection to be idle when a breakpoint was hit and disconnect after 5 seconds. This PR disables HTTP Keep-Alive and sets the socket timeout to 0 so that breakpoints won't cause the connection to timeout.

Fixes https://github.com/firebase/firebase-tools/issues/5991.