Reconcile existing secrets management code.
Updated the original grant access code to accept a secret and a MultiServiceAccounts
so that it can be used to grant access to multiple accounts at the same time. Logic
for verifying a secret exists or getting a backend's service account are trivial
and were moved into the command.

Updated the original secrets dialog to use the new MultiServiceAccounts type so that
we can remember which accounts need the special access for Cloud Build to pin versions.