Re-allow users to not select an emulator UI port. (#8626)
* Re-allow users to not select an emulator UI port.

Fixes #8621

Sorry; somehow I missed that there was a case where numbers are not required.
Keeps the library's behavior of assuming that numbers are required, but allows
explicitly overwriting this with a "reququired: false". This is the only
value that can be provided to "required". Using type magic, the return type
of number can only be defined if "required: false" is passed in.

* Format & changelog

---------

Co-authored-by: joehan <joehanley@google.com>