feat: Add support for function prefixes (#8911)
This change adds a new `prefix` property to firebase functions configuration in firebase.json. The new property can be used to prepend function name with the given prefix for all functions in the codebase.

Note that prefix is also used for all secret associated with the codebase to allow different codebases to refer to different configuration value.

The new feature also requires two new validation rules for function configurations in `firebase.json`:

1.  The `prefix` property, if specified, must only contain lowercase letters, numbers, and hyphens.
2.  The combination of `source` and `prefix` must be unique across all function configurations.

These changes ensure that function deployments with prefixes are valid and that there are no conflicts when deploying multiple functions from the same source directory.