fix: index-spec matching for optional values and vectorConfig.
The few new optional values for indexes were added recently via
https://github.com/firebase/firebase-tools/pull/8939.

When comparing the local index with the server index, the default value of optional index configurations should be taken into account depending on the database edition (standard or enterprise).

I also found that the matching logic for vectorConfig was not working correctly, so I fixed that as well.

Fixes #8859.