Add new apphosting:env:import commmand
Approved by committee ages ago. This will expand to remote env by just
adding a --backend flag and can also include a dialog if neither backend
nor apphosting.yaml is found.

Only controversial change is that I updated the discoverBackendRoot to
also look for .env files. This seems to also be a pretty good indication
of the root of a website and helps make sure that this command is
always operating on the same directory.